// src/types/grpc/proto/client/wallet.ts
/**
* This file is auto-generated by nestjs-proto-gen-ts
*/

import { Observable } from 'rxjs';
import { Metadata } from '@grpc/grpc-js';

export namespace wallet {
    export interface WalletService {
        createWallet(
            data: CreateWalletRequest,
            metadata?: Metadata,
            ...rest: any[]
        ): Observable<CreateWalletResponse>;
    }
    export interface CreateWalletRequest {
        mobile?: string;
    }
    export interface CreateWalletResponse {
        message?: string;
        data?: wallet.Wallet;
        status?: number;
    }
    export interface Wallet {
        id?: string;
        mobile?: string;
        balance?: number;
        userId?: string;
        currency?: string;
        status?: string;
        createdAt?: number;
        updatedAt?: number;
    }
}